LayoutTests/http/tests/adClickAttribution/store-ad-click-attribution-expected.txt

Tests storage of ad click attribution.


WebCore::AdClickAttribution 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 3
No conversion data.
